dashboard/multiuser/requirements.txt
# Multi-User SaaS Platform Additional Requirements

# Authentication
bcrypt>=4.0.1
streamlit-authenticator>=0.2.3

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
sqlalchemy>=2.0.23      # ORM (optional, for advanced usage)

# Indian Broker APIs
kiteconnect>=4.2.0      # Zerodha Kite API
upstox-python-sdk>=2.5.0  # Upstox API

# Background Jobs (optional - for autonomous trading)
celery>=5.3.4
redis>=5.0.1

# Multi-user features
python-jose[cryptography]>=3.3.0  # JWT tokens (optional)
passlib>=1.7.4                     # Password hashing utilities

# Already in main requirements.txt:
# - streamlit
# - plotly
# - pandas
# - alpaca-py (US broker)
# - python-dotenv
